Fix leaking temp dir
